+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * New fsio_copy_file functions.
 * New fsio_path_exists function.
 * New fsio_file_exists function.
+* New fsio_create_empty_file function.
 * Free all memory in tests
 * Test memory leaks in CI